Add -kK for mkrc
`mkrc` is implemented internally by calling the `rcup` command with
certain flags. This has the (surprising?) effect of calling the pre-up
and post-up hooks for each file passed to `mkrc`, which may be
undesirable.

The hooks now default to being skipped when running `mkrc`. They can be
run using `-k`, and if needed forced using `-K`.